Restructure Spotify script under scripts/spotify-flow
