+++ package.json
@@ -4,10 +4,10 @@
   "version": "0.1.0",
   "private": true,
   "description": "Spotify liked songs export script with TS/ESLint/Prettier tooling",
-  "main": "dist/index.js",
+  "main": "dist/scripts/spotify-flow/index.js",
   "type": "commonjs",
   "scripts": {
-    "start": "ts-node src/index.ts",
+    "start": "ts-node src/scripts/spotify-flow/index.ts",
     "build": "tsc -p tsconfig.json",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "lint": "eslint . --ext .ts",
